requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
scipy>=1.7.0
tokenizers>=0.13.0
datasets>=2.10.0
accelerate>=0.20.0
einops>=0.6.0
tqdm>=4.64.0
wandb>=0.15.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
pandas>=1.5.0
pyyaml>=6.0
